// backend/src/enums/NotificationTitle.ts
enum NotificationTitle {
  ADD_PRODUCT = "add_product",
  UPDATE_PRODUCT = "update_product",
  DELETE_PRODUCT = "delete_product",
  SECTION_ITEMS_UPDATED = "section_items_updated",
  ADD_ADMIN_BY_MANAGER = "add_admin_by_manager",
  DELETE_ADMIN_BY_MANAGER = "delete_admin_by_manager",
  SUSPEND_ADMIN_BY_MANAGER = "suspend_admin_by_manager",
  UPGRADE_ADMIN_BY_MANAGER = "upgrade_admin_by_manager",
  DOWNGRADE_ADMIN_BY_MANAGER = "downgrade_admin_by_manager",
  UPDATE_PRODUCT_IMAGE = "update_product_image",
  DELETE_PRODUCT_IMAGE = "delete_product_image",
  REMOVE_PRODUCT_FROM_SECTION = "remove_product_from_section",
  ACTIVE_ADMIN_BY_MANAGER = "active_admin_by_manager",
}

export default NotificationTitle;